dapr_multi_agent_actor_example/requirements.txt
dapr-agents>=0.1.0
python-dotenv
requests
fastapi
uvicorn
httpx